Refactor ManagePage form into own component
